Update DMS_Titles.xml (#76)
Okay this one was a bit of a headache to me, so I have reworked the lore here: I think that Pre-sergeants are outsiders given rank, but are not recognized a full citizens of the Colonial faction until they get the Reservist rank. 

Then the proper ranking would be: Private to Corporal to Sergeant as NCOs or non-commissioned officers, the last two are backbone of armies as they are in the frontline and the sergeant serves as the connection between officers and enlisted. 

Warrant Officers are specialist troops, but can serve as a stepping stone for enlisted soldiers and sergeants who wish to become officers.

Then it's Lieutenant to Captain to Major to Colonel and finally the highest rank of Brigadier.